Move resource model reader to new bundle.
