+++ extensions/modelreader/src/test/resources/serialization/resourceModelWrapper.xml
